Added Model Class AlumniModel
The AlumniModel class represents an individual alumni entity in the application and serves as a blueprint for storing and retrieving information related to alumni.
The class includes several properties, each with a detailed description.
'Id' represents the unique identifier for alumni and its type is int
'Name' stores the name of alumni
'GraduationYear' stores the date of graduation
'ContactDetails' stores the contact information of the alumni
'FacebookHandle' stores the facebook profile url
'LinkedInHandle' stores the info of linked in url
'TwitterHandle' stores the twiter profile url
'Email' stores the email
'CurrentPosition' stores the position of alumni the company
'CurrentEmployer' stores the current employer name